iret-to-v86: only 16-bit eip, thank you.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@60881 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/IEMAllCImpl.cpp.h
@@ -2923,8 +2923,8 @@ IEM_CIMPL_DEF_5(iemCImpl_iret_prot_v8086, PCPUMCTX, pCtx, uint32_t, uNewEip, uin
     iemCImplCommonV8086LoadSeg(&pCtx->ds, uNewDs);
     iemCImplCommonV8086LoadSeg(&pCtx->fs, uNewFs);
     iemCImplCommonV8086LoadSeg(&pCtx->gs, uNewGs);
-    pCtx->rip      = uNewEip;
-    pCtx->rsp      = uNewEsp;
+    pCtx->rip      = (uint16_t)uNewEip;
+    pCtx->rsp      = uNewEsp; /** @todo check this out! */
     pIemCpu->uCpl  = 3;
 
     return VINF_SUCCESS;
